Small changes to attend py 3.10, fix datasources acquirement and parsing, fix dead lock postgree on inserting new molecules

+++ package/chemicalchecker/util/download/download.py
@@ -3,6 +3,7 @@ import os
 import sys
 import shutil
 import tempfile
+import zlib
 from glob import glob
 from time import sleep
 from ftplib import FTP
@@ -15,12 +16,14 @@ from chemicalchecker.util import logged
 from chemicalchecker.util import Config
 from chemicalchecker.util import psql
 
+import requests as reqs
+from requests.auth import HTTPBasicAuth
 
 @logged
 class Downloader():
     """Downloader class."""
 
-    def __init__(self, url, data_path, tmp_dir=None, dbname=None, file=None):
+    def __init__(self, url, data_path, tmp_dir=None, dbname=None, file=None, username=None, password=None):
         """Initialize a Download instance.
 
         Download from url, unpack in tmp_dir, and copy to data_path.
@@ -36,8 +39,11 @@ class Downloader():
         self.data_path = data_path
         self.dbname = dbname
         self.file = file
+        self.username = username
+        self.password = password
         if not tmp_dir:
-            tmp_dir = Config().PATH.CC_TMP
+            cc_config = kwargs.get("cc_config", os.environ['CC_CONFIG'])
+            tmp_dir = Config(cc_config).PATH.CC_TMP
         self.tmp_dir = tmp_dir
         try:
             self.url = Downloader.validate_url(url)
@@ -95,8 +101,26 @@ class Downloader():
             raise RuntimeError('Unrecognized URL protocol.')
         return new_url
 
+    
+
+    def _download_and_decompress_file( self, url, filename):
+        response = reqs.get(url, stream=True)
+        decompressor = zlib.decompressobj(16 + zlib.MAX_WBITS)
+        
+        with open(filename, 'w+') as f:
+            while True:
+                try:
+                    chunk = response.raw.read(1024)
+                    if not chunk:
+                        break
+                    string = decompressor.decompress(chunk).decode("utf-8",errors='ignore')
+                    f.write(string)
+                except:
+                    pass
+                
     def download(self):
         """Perform the download."""
+        print( f"Downloading datasource { self.url }" )
         try:
             import wget
         except ImportError:
@@ -113,33 +137,56 @@ class Downloader():
         self.__log.debug('temp download dir %s', tmp_dir)
         # set wget user agent
 
-        class MyOpener(urllib.request.FancyURLopener):
-            version = 'Mozilla/5.0'
-        wget.ulib.urlretrieve = MyOpener().retrieve
-        # determine file name
-        parsed = urlparse(self.url)
-        tmp_file = os.path.join(tmp_dir, wget.detect_filename(self.url))
-        # download
-        if parsed.scheme == 'file':
-            # file can be on local filesystem
-            shutil.copy(parsed.path, tmp_dir)
-        else:
-            # or has to be downloaded, in case try several times
-            attempts = 0
-            downloaded = False
-            while attempts < 5:
-                try:
-                    wget.download(self.url, tmp_file)
-                    downloaded = True
-                    break
-                except Exception as err:
-                    attempts += 1
-                    self.__log.warning('Attempt failed: %s', str(err))
-                    request.urlcleanup()
-                    sleep(5)
-            if not downloaded:
-                raise Exception('All attempts to download failed.')
-        # unzip
+        if(self.username!=None and self.password!=None):
+            tmp_file = os.path.join( tmp_dir, 'dbfile.xml.zip')
+            try:
+                response = reqs.get( self.url, auth = HTTPBasicAuth( self.username, self.password ), stream=True )
+                with open(tmp_file, 'wb') as zfile:
+                    for i, chunk in enumerate(response.iter_content(chunk_size=1024)):
+                        zfile.write(chunk)
+            except:
+                raise Exception('Failed on authenticated data acquirement')
+        else:            
+            class MyOpener(urllib.request.FancyURLopener):
+                version = 'Mozilla/5.0'
+            wget.ulib.urlretrieve = MyOpener().retrieve
+            tpfname = wget.detect_filename( self.url )
+            if( tpfname == 'download.wget' ):
+                if( self.file != None ):
+                    tpfname = self.file
+                else:
+                    raise Exception('Cannot guess the filename from the url neither a filename was given as input')
+            # determine file name
+            parsed = urlparse(self.url)
+            tmp_file = os.path.join(tmp_dir, tpfname )
+            # download
+            if parsed.scheme == 'file':
+                # file can be on local filesystem
+                shutil.copy(parsed.path, tmp_dir)
+            else:
+                # or has to be downloaded, in case try several times
+                attempts = 0
+                downloaded = False
+                while attempts < 5:
+                    try:
+                        wget.download(self.url, tmp_file)
+                        downloaded = True
+                        break
+                    except Exception as err:
+                        attempts += 1
+                        self.__log.warning('Attempt failed: %s', str(err))
+                        request.urlcleanup()
+                        sleep(5)
+                if not downloaded:
+                    if( self.url.endswith('.gz') ):
+                        if not os.path.exists(self.data_path):
+                            os.makedirs(self.data_path, 0o775)
+                        destination = os.path.join( self.data_path, self.file)
+                        self._download_and_decompress_file( self.url, destination )
+                    else:
+                        raise Exception('All attempts to download failed.')
+
+        # Decompressing pssible zip files
         tmp_unzip_dir = os.path.join(tmp_dir, 'unzip')
         os.mkdir(tmp_unzip_dir)
         self.__log.debug('temp unzip dir %s', tmp_unzip_dir)
@@ -156,6 +203,8 @@ class Downloader():
                 self.__log.error('problem uncompressing %s', tmp_file)
                 self.__log.error('error was: %s', str(err))
                 raise err
+
+        # Treating database sql files                    
         if self.dbname is not None:
             file_path = os.path.join(tmp_unzip_dir, self.file)
             if '*' in self.file:
@@ -163,6 +212,9 @@ class Downloader():
                 paths = glob(file_path)
                 if len(paths) > 1:
                     raise Exception("`*` in %s db_file is ambigous.", self)
+                if len(paths) == 0:
+                    raise Exception("There are no paths in the zipped file.", self)
+                    
                 file_path = paths[0]
             cmd2run = 'PGPASSWORD=' + Config().DB.password + \
                 ' dropdb --if-exists -h ' + Config().DB.host + \
@@ -172,11 +224,17 @@ class Downloader():
                 " createdb -h " + Config().DB.host + \
                 " -U " + Config().DB.user + \
                 ' ' + self.dbname + " && "
-
-            cmd2run += 'PGPASSWORD=' + Config().DB.password + \
-                ' psql -h ' + Config().DB.host + \
-                " -U " + Config().DB.user + \
-                ' -d ' + self.dbname + ' <' + file_path
+                
+            if( file_path.endswith('.dmp') ):
+                cmd2run += 'PGPASSWORD=' + Config().DB.password + \
+                ' pg_restore --no-privileges --no-owner -h ' + Config().DB.host+ \
+                " -U " + Config().DB.user+ \
+                ' -d ' + self.dbname + ' < ' + file_path
+            else:
+                cmd2run += 'PGPASSWORD=' + Config().DB.password + \
+                ' psql -h ' + Config().DB.host+ \
+                " -U " + Config().DB.user+ \
+                ' -d ' + self.dbname + ' < ' + file_path
 
             # cmd2run += 'PGPASSWORD=' + Config().DB.password +
             # ' pg_restore -h ' + Config().DB.host +
@@ -215,7 +273,9 @@ class Downloader():
                 os.makedirs(self.data_path, 0o775)
             destination = os.path.join(self.data_path, self.file)
 
-        shutil.move(source, destination)
+        if( not os.path.isfile(destination) ):
+            shutil.move(source, destination)
+    
         self.__log.debug('downloaded to  %s', self.data_path)
         # remove temp dir
         shutil.rmtree(tmp_dir)
